move id mapping code out of channel
